+++ demo/RadiogroupDemo.jsx
@@ -7,10 +7,13 @@ const { Item } = RadioGroup;
 class Demo extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {};
+    this.state = {
+      value: 'apple',
+    };
   }
 
   handleChange(value) {
+    console.log(`value is changed, new value is ${value}`);
     const me = this;
     me.setState({
       value,
@@ -20,8 +23,8 @@ class Demo extends React.Component {
   render() {
     const me = this;
     return (
-      <RadioGroup value={me.state.value} onChange={me.handleChange.bind(me)} disabled>
-        <Item disabled={false} value="apple" text="苹果" />
+      <RadioGroup value={me.state.value} onChange={me.handleChange.bind(me)}>
+        <Item value="apple" text="苹果" />
         <Item value="banana" text="香蕉" />
         <Item value="orange" text="橘子" />
       </RadioGroup>